Changing the return type for a curl function to get rid of a warning caused by newer versions of libcurl.

+++ src/util.c
@@ -5097,7 +5097,7 @@ struct FtpFile {
 
 
 
-int curl_fwrite(void *buffer, size_t size, size_t nmemb, void *stream) {
+size_t curl_fwrite(void *buffer, size_t size, size_t nmemb, void *stream) {
   struct FtpFile *out = (struct FtpFile *)stream;
   if (out && !out->stream) {
     out->stream=fopen(out->filename, "wb");
